Basic support for C++11, using ghc std::filesystem compatibility library.

+++ examples/hscpp-example-utils/include/hscpp-example-utils/MemoryManager.h
@@ -21,7 +21,7 @@ private:
     {
         bool bFree = false;
         bool bExternallyOwned = false;
-        size_t capacity = 0;
+        uint64_t capacity = 0;
         uint8_t* pMemory = nullptr;
     };
 
@@ -74,14 +74,14 @@ public:
     {
         if (ref.id < m_Blocks.size())
         {
-            Block& block = m_Blocks.at(ref.id);
+            Block& block = m_Blocks.at(static_cast<size_t>(ref.id));
             block.bFree = true;
 
             T* pObject = reinterpret_cast<T*>(block.pMemory);
             pObject->~T();
 
             // Zero out memory to cause crashes when dereferencing dangling pointers.
-            std::memset(block.pMemory, 0, block.capacity);
+            std::memset(block.pMemory, 0, static_cast<size_t>(block.capacity));
         }
     }
 
@@ -103,7 +103,7 @@ public:
         return ref;
     }
 
-    uint8_t* GetMemory(size_t id) override;
+    uint8_t* GetMemory(uint64_t id) override;
 
     //============================================================================
     // hscpp::IAllocator implementation.
@@ -113,7 +113,7 @@ public:
     uint64_t Hscpp_Free(uint8_t* pMemory) override;
 
 private:
-    size_t TakeFirstFreeBlock(size_t size);
+    size_t TakeFirstFreeBlock(uint64_t size);
 
     std::vector<Block> m_Blocks;
     hscpp::AllocationResolver* m_pHscppAllocationResolver = nullptr;
